Update sldemo_househeatExample.m
This example demonstrates how to create a thermal model of a house using Simulink(R). The model consists of three main components: the outdoor environment, the thermal characteristics of the house, and the house heating system.

The model is initialized with data from the "sldemo_househeat_data.m" file which specifies the geometry of the house, the thermal properties of the materials used in its construction, the thermal resistance of the house, the heater characteristics, the cost of electricity, and the initial room temperature.

The time unit in the model is hours, and certain quantities like air flow rate are expressed per hour (not per second).

Model Components
Set Point
The Set Point is a constant block that specifies the temperature that must be maintained indoors. The default temperature is 70 ºF, but the model converts it to ºC.

Thermostat
The Thermostat subsystem contains a Relay block that allows temperature fluctuations of 5 ºF above or below the desired room temperature. If the air temperature drops below 65 ºF, the thermostat turns on the heater.

Heater
The Heater subsystem models a constant air flow rate, Mdot, which is specified in the "sldemo_househeat_data.m" file. The thermostat signal turns the heater on or off. When the heater is on, it blows hot air at a constant flow rate of Mdot (1kg/sec = 3600kg/hr by default) and temperature THeater (50 ºC = 122 ºF by default). Equation 1 expresses the heat flow into the room.

Equation 1:

$$\frac{dQ}{dt}=\left( T_{heater} - T_{room} \right) \cdot Mdot \cdot c$$

where:

$$\frac{dQ}{dt} = \mbox{ heat flow from the heater into the room}$$

$$c = \mbox{ heat capacity of air at constant pressure}$$

$$Mdot = \mbox{ air mass flow rate through heater (kg/hr)}$$

$$T_{heater} = \mbox{ temperature of hot air from heater}$$

$$T_{room} = \mbox{ current room air temperature}$$

Cost Calculator
The Cost Calculator is a Gain block that integrates the heat flow over time and multiplies it by the energy cost. The model plots the heating cost in the PlotResults scope.

House
The House subsystem calculates room temperature variations, taking into consideration the heat flow from the heater and heat losses to the environment. Heat losses and the temperature time derivative are expressed by Equation 2.

Equation 2:

$$\left( \frac{dQ}{dt} \right) {losses} = \frac{T{room}-T_{out}}{R_{eq}}$$

$$\frac{dT_{room}}{dt} = \frac{1}{M_{air} \cdot c} \cdot \left( \frac{dQ_{heater}}{dt} - \frac{dQ_{losses}}{dt} \right)$$

where:

$$\left( \frac{dQ}{dt} \right) _{losses} = \mbox{ heat loss rate}$$

$$T_{out} = \mbox{ outdoor temperature}$$

$$R_{eq} = \mbox{ thermal resistance of the house}$$

$$\frac{dT_{room}}{dt} = \mbox{ temperature time derivative}$$